fix scripts

+++ scripts/link_dots.sh
@@ -1,6 +1,22 @@
 #!/bin/bash
 
+# This script generates symbolic links to dot files in this directory.
+
 this_dir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
-ln -s $this_dir/vimrc $HOME/.vimrc
-ln -s $this_dir/gvimrc $HOME/.gvimrc
+# .vimrc
+if ! [ -L $HOME/.vimrc ]; then
+  echo "linking .vimrc"
+  ln -s $this_dir/vimrc $HOME/.vimrc
+else
+  echo ".vimrc is already linked."  
+fi
+
+# .gvimrc
+if ! [ -L $HOME/.gvimrc ]; then
+  echo "linking .vimrc"
+  ln -s $this_dir/gvimrc $HOME/.gvimrc
+else
+  echo ".gvimrc is already linked."  
+fi
+
